# Improve performance of sdssplitargs