--- a/src/unit/test_sds.c
+++ b/src/unit/test_sds.c
@@ -1,6 +1,8 @@
+#include <ctype.h>
 #include <stdio.h>
 #include <limits.h>
 #include <string.h>
+#include <sys/time.h>
 #include "test_help.h"
 
 #include "../sds.h"
@@ -325,6 +327,255 @@ int test_sdsHeaderSizes(int argc, char **argv, int flags) {
     TEST_ASSERT_MESSAGE("can't always adjust SDS_TYPE_32 with SDS_TYPE_64",
                         sizeof(struct sdshdr64) <= 2 * sizeof(struct sdshdr32) + 1);
 #endif
+
+    return 0;
+}
+
+int test_sdssplitargs(int argc, char **argv, int flags) {
+    UNUSED(argc);
+    UNUSED(argv);
+    UNUSED(flags);
+
+    int len;
+    sds *sargv;
+
+    sargv = sdssplitargs("Testing one two three", &len);
+    TEST_ASSERT(4 == len);
+    TEST_ASSERT(!strcmp("Testing", sargv[0]));
+    TEST_ASSERT(!strcmp("one", sargv[1]));
+    TEST_ASSERT(!strcmp("two", sargv[2]));
+    TEST_ASSERT(!strcmp("three", sargv[3]));
+    sdsfreesplitres(sargv, len);
+
+    sargv = sdssplitargs("", &len);
+    TEST_ASSERT(0 == len);
+    TEST_ASSERT(sargv != NULL);
+    sdsfreesplitres(sargv, len);
+
+    sargv = sdssplitargs("\"Testing split strings\" \'Another split string\'", &len);
+    TEST_ASSERT(2 == len);
+    TEST_ASSERT(!strcmp("Testing split strings", sargv[0]));
+    TEST_ASSERT(!strcmp("Another split string", sargv[1]));
+    sdsfreesplitres(sargv, len);
+
+    sargv = sdssplitargs("\"Hello\" ", &len);
+    TEST_ASSERT(1 == len);
+    TEST_ASSERT(!strcmp("Hello", sargv[0]));
+    sdsfreesplitres(sargv, len);
+
+    return 0;
+}
+
+/* Included here is the legacy implementation of sdssplitargs to do fuzzing against. */
+
+/* Helper function for sdssplitargs() that returns non zero if 'c'
+ * is a valid hex digit. */
+static int is_hex_digit(char c) {
+    return (c >= '0' && c <= '9') || (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F');
+}
+
+/* Helper function for sdssplitargs() that converts a hex digit into an
+ * integer from 0 to 15 */
+static int hex_digit_to_int(char c) {
+    switch (c) {
+    case '0': return 0;
+    case '1': return 1;
+    case '2': return 2;
+    case '3': return 3;
+    case '4': return 4;
+    case '5': return 5;
+    case '6': return 6;
+    case '7': return 7;
+    case '8': return 8;
+    case '9': return 9;
+    case 'a':
+    case 'A': return 10;
+    case 'b':
+    case 'B': return 11;
+    case 'c':
+    case 'C': return 12;
+    case 'd':
+    case 'D': return 13;
+    case 'e':
+    case 'E': return 14;
+    case 'f':
+    case 'F': return 15;
+    default: return 0;
+    }
+}
+
+static sds *sdssplitargs_legacy(const char *line, int *argc) {
+    const char *p = line;
+    char *current = NULL;
+    char **vector = NULL;
+
+    *argc = 0;
+    while (1) {
+        /* skip blanks */
+        while (*p && isspace(*p)) p++;
+        if (*p) {
+            /* get a token */
+            int inq = 0;  /* set to 1 if we are in "quotes" */
+            int insq = 0; /* set to 1 if we are in 'single quotes' */
+            int done = 0;
+
+            if (current == NULL) current = sdsempty();
+            while (!done) {
+                if (inq) {
+                    if (*p == '\\' && *(p + 1) == 'x' && is_hex_digit(*(p + 2)) && is_hex_digit(*(p + 3))) {
+                        unsigned char byte;
+
+                        byte = (hex_digit_to_int(*(p + 2)) * 16) + hex_digit_to_int(*(p + 3));
+                        current = sdscatlen(current, (char *)&byte, 1);
+                        p += 3;
+                    } else if (*p == '\\' && *(p + 1)) {
+                        char c;
+
+                        p++;
+                        switch (*p) {
+                        case 'n': c = '\n'; break;
+                        case 'r': c = '\r'; break;
+                        case 't': c = '\t'; break;
+                        case 'b': c = '\b'; break;
+                        case 'a': c = '\a'; break;
+                        default: c = *p; break;
+                        }
+                        current = sdscatlen(current, &c, 1);
+                    } else if (*p == '"') {
+                        /* closing quote must be followed by a space or
+                         * nothing at all. */
+                        if (*(p + 1) && !isspace(*(p + 1))) goto err;
+                        done = 1;
+                    } else if (!*p) {
+                        /* unterminated quotes */
+                        goto err;
+                    } else {
+                        current = sdscatlen(current, p, 1);
+                    }
+                } else if (insq) {
+                    if (*p == '\\' && *(p + 1) == '\'') {
+                        p++;
+                        current = sdscatlen(current, "'", 1);
+                    } else if (*p == '\'') {
+                        /* closing quote must be followed by a space or
+                         * nothing at all. */
+                        if (*(p + 1) && !isspace(*(p + 1))) goto err;
+                        done = 1;
+                    } else if (!*p) {
+                        /* unterminated quotes */
+                        goto err;
+                    } else {
+                        current = sdscatlen(current, p, 1);
+                    }
+                } else {
+                    switch (*p) {
+                    case ' ':
+                    case '\n':
+                    case '\r':
+                    case '\t':
+                    case '\0': done = 1; break;
+                    case '"': inq = 1; break;
+                    case '\'': insq = 1; break;
+                    default: current = sdscatlen(current, p, 1); break;
+                    }
+                }
+                if (*p) p++;
+            }
+            /* add the token to the vector */
+            vector = s_realloc(vector, ((*argc) + 1) * sizeof(char *));
+            vector[*argc] = current;
+            (*argc)++;
+            current = NULL;
+        } else {
+            /* Even on empty input string return something not NULL. */
+            if (vector == NULL) vector = s_malloc(sizeof(void *));
+            return vector;
+        }
+    }
+
+err:
+    while ((*argc)--) sdsfree(vector[*argc]);
+    s_free(vector);
+    if (current) sdsfree(current);
+    *argc = 0;
+    return NULL;
+}
+
+static long long timeInMicroseconds(void) {
+    struct timeval tv;
+
+    gettimeofday(&tv, NULL);
+    return (((long long)tv.tv_sec) * 1000000) + (tv.tv_usec);
+}
+
+#define PRINTABLE_ASCII_START 32
+#define PRINTABLE_ASCII_END 126
+
+static char get_random_printable_char(void) {
+    return (char)(PRINTABLE_ASCII_START + rand() % (PRINTABLE_ASCII_END - PRINTABLE_ASCII_START + 1));
+}
+
+int test_sdssplitargs_benchmark(int argc, char **argv, int flags) {
+    if (!(flags & UNIT_TEST_ACCURATE)) return 0;
+
+    UNUSED(argc);
+    UNUSED(argv);
+
+    /* Seed the random number generator */
+    unsigned int seed = time(NULL);
+    srand(seed);
+    TEST_PRINT_INFO("Using random seed: %d", seed);
+
+    for (int j = 0; j < 1000; j++) {
+        long i = 0;
+        sds random = sdsnew("");
+        char bytes[] = "X";
+        long long start, elapsed_new, elapsed_old;
+        for (int j = 0; j < 100; j++){
+            bytes[0] = get_random_printable_char();
+            random = sdscatlen(random, bytes, 1);
+        }
+
+        /* Validate the new and legacy implementation return logically the 
+         * same value. */
+        int len1, len2;
+        sds *argv1 = sdssplitargs(random, &len1);
+        sds *argv2 = sdssplitargs_legacy(random, &len2);
+
+        TEST_ASSERT(len1 == len2);
+        for (int j = 0; j < len1; j++) {
+            TEST_ASSERT(!sdscmp(argv1[j], argv2[j]));
+        }
+
+        int valid_result = argv1 != NULL;
+        sdsfreesplitres(argv1, len1);
+        sdsfreesplitres(argv2, len2);
+
+        /* Only benchmark valid strings */
+        if (!valid_result) continue;
+
+        const int iterations = 10000;
+        start = timeInMicroseconds();
+        for (; i < iterations; i++) {
+            int len;
+            sds *argv1 = sdssplitargs(random, &len);
+            sdsfreesplitres(argv1, len);
+        }
+        elapsed_new = timeInMicroseconds() - start;
+
+        start = timeInMicroseconds();
+        for (i = 0; i < iterations; i++) {
+            int len;
+            sds *argv1 = sdssplitargs_legacy(random, &len);
+            sdsfreesplitres(argv1, len);
+        }
+        elapsed_old = timeInMicroseconds() - start;
+        TEST_PRINT_INFO("Improvement: %.2f%%, new:%lldus, old:%lldus",
+                        (1.0 - (float)elapsed_new/elapsed_old) * 100,
+                        elapsed_new,
+                        elapsed_old);
+        sdsfree(random);
+    }
 
     return 0;
 }