@madolson 2024-10-29
> Is sdssplitargs used in any hot code path?

It's used in the inline protocol command parsing, which isn't super commonly used but I've seen it in production. It's also in some command parsing like configs.